refactor: user-activity
